mp4 code cleanup.
Clarify lifetime of Track objects using shared_ptr.  This is because
the demux stream and seek table both are reference counted and could
theoretically outlive one another.

Also get rid of move(mp4) into the demux stream.  The idea was to
avoid copying vectors but the stream really only needs the mdat
offsets (2 integers) and the track (which is now reference counted).

Lastly get rid of DropSeekTable(), make the seek table a singleton
and have its destructor clear out that memory.  I am not sure how
big stts is in practice and whether any of that has a point, but
leave it in.