mp3 & aac: Place cap on decoder retries, eventually treat errors as EOF.
Two scenarios to account for:

1. Bad codec data at the end of the file, but valid frame headers in tact.
I have an MP3 like this.  Too many retries makes this code not make progress
towards discovering we've hit EOF and just makes it complain about bad decodes
all day long.  For this we cap the number of retries.

2. Last frame is invalid and seek to next one fails.  This wouldn't cause a
tight loop of failure like #1, but it would return an error to Read() potentially
after a valid decode, causing us to drop some good samples.  For this we turn
decoder errors into EOF.