third_party/Makefile.inc: re-arrange some deck chairs for Windows build.
The reason the source file lists are so split up like this is because some
commands generated by make run into the Windows NT limit for a command
line length if all filenames are concatenated together.  I bumped into
another one of these limits in another script, so adjust where the split
is made.