Refactor classic unix-style device's enumerators
Fold in some common parts of oss and /dev/audio drivers.

Firstly, the enumerator implementations will actually return all
devices [unlike the old one which only considered a default device].

In the presence of multiple devices we won't open(2) all of them
immediately, we will list /dev to figure out what nodes are there
and open them when the device object is requested.

Lastly, introduce some notion of mixer devices.  Check the MIXERDEV
environment variable.  If not present, try to intuit the correct one
from AUDIODEV if that is specified.  This was the impetus for the
whole refactor, I want to support mixer devices.