Introduce mute API.
For a few device backends, try to use the native platform mute flags.
For others, we have a "soft mute" which will just set volume to 0.

Some device backends probably need more work.  And the "soft mute" is
kind of crude.  But seems to work well with Mac, ALSA, and Windows,
which is a good chunk of relevance.